Move maneuver_path to DreadnoughtPlayView
