+++ lib/chukinas_web/views/dreadnought.ex
@@ -10,17 +10,6 @@ defmodule ChukinasWeb.DreadnoughtView do
   alias Chukinas.Util.Opts
 
 
-  def maneuver_path(%Ev.Maneuver{} = path, unit_id) do
-    assigns =
-      path
-      |> Map.from_struct
-      |> Map.put(:unit_el_id, "unit-#{unit_id}")
-    render("maneuver_path.html", assigns)
-  end
-  def maneuver_path(_, _) do
-    nil
-  end
-
   def unit_event(%Ev.Maneuver{} = event) do
     standard_attributes(event, "maneuver") ++ [
       data_attr("id", event.id),
